[FIX] Fix the issue that game object did not clear completely when doing scene reload. [MOD] the audio volume modify way.

+++ Game1007Project/Game1007Project/include/audioPlayer.h
@@ -6,10 +6,10 @@
 class AudioPlayer
 {
 public:
-	void PlaySound(int channel, std::string sound, float volume);
-	void PlaySound(int channel, std::string sound, int times, float volume);
-	void PlaySound(int channel, std::string sound, std::function<void(void)> callback, float volume);
-	void PlaySound(int channel, std::string sound, int times, std::function<void(void)> callback, float volume);
+	void PlaySound(int channel, std::string sound, int volume);
+	void PlaySound(int channel, std::string sound, int times, int volume);
+	void PlaySound(int channel, std::string sound, std::function<void(void)> callback, int volume);
+	void PlaySound(int channel, std::string sound, int times, std::function<void(void)> callback, int volume);
 	void PlayMusic(std::string music);
 	static AudioPlayer* GetInstance();
 private:
